export meta data validation
