+++ packages/xrpl/src/models/transactions/vaultCreate.ts
@@ -14,9 +14,10 @@ import {
   VAULT_DATA_MAX_BYTE_LENGTH,
   XRPLNumber,
   isXRPLNumber,
+  MAX_MPT_META_BYTE_LENGTH,
+  getValidationMessagesForMPTokenMetadata,
+  MPT_META_WARNING_HEADER,
 } from './common'
-
-const META_MAX_BYTE_LENGTH = 1024
 
 /**
  * Enum representing withdrawal strategies for a Vault.
@@ -85,13 +86,14 @@ export interface VaultCreate extends BaseTransaction {
   DomainID?: string
 }
 
+/* eslint-disable max-lines-per-function -- Not needed to reduce function */
+/* eslint-disable max-statements -- required to do all field validations */
 /**
  * Verify the form and type of an {@link VaultCreate} at runtime.
  *
  * @param tx - A {@link VaultCreate} Transaction.
  * @throws When the {@link VaultCreate} is malformed.
  */
-// eslint-disable-next-line max-lines-per-function -- required to do all field validations
 export function validateVaultCreate(tx: Record<string, unknown>): void {
   validateBaseTransaction(tx)
 
@@ -123,9 +125,9 @@ export function validateVaultCreate(tx: Record<string, unknown>): void {
       )
     }
     const metaByteLength = metaHex.length / 2
-    if (metaByteLength > META_MAX_BYTE_LENGTH) {
+    if (metaByteLength > MAX_MPT_META_BYTE_LENGTH) {
       throw new ValidationError(
-        `VaultCreate: MPTokenMetadata exceeds ${META_MAX_BYTE_LENGTH} bytes (actual: ${metaByteLength})`,
+        `VaultCreate: MPTokenMetadata exceeds ${MAX_MPT_META_BYTE_LENGTH} bytes (actual: ${metaByteLength})`,
       )
     }
   }
@@ -139,4 +141,20 @@ export function validateVaultCreate(tx: Record<string, unknown>): void {
       'VaultCreate: Cannot set DomainID unless tfVaultPrivate flag is set.',
     )
   }
+
+  if (tx.MPTokenMetadata != null) {
+    const result = getValidationMessagesForMPTokenMetadata(tx.MPTokenMetadata)
+
+    if (!result.isValid) {
+      const message = [
+        MPT_META_WARNING_HEADER,
+        ...result.validationMessages,
+      ].join('\n')
+
+      // eslint-disable-next-line no-console -- Required here.
+      console.warn(message)
+    }
+  }
 }
+/* eslint-enable max-lines-per-function */
+/* eslint-enable max-statements */
